% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMI_on_AgingClock.R
\name{Impact_of_imputation_on_horvath_clock}
\alias{Impact_of_imputation_on_horvath_clock}
\title{A function to show the impact of missing values imputation methods on aging clock}
\usage{
Impact_of_imputation_on_horvath_clock(extendedBetaData, metaData, coefHorvath)
}
\arguments{
\item{extendedBetaData}{An expression data frame containing CpGs in rows and samples and  CpG position in columns}

\item{metaData}{Meta data frame samples informations}

\item{coefHorvath}{Horvath reference file}
}
\value{
A list of data frame with estimated ages
}
\description{
A function to show the impact of missing values imputation methods on aging clock
}
\examples{
Impact_of_imputation_on_horvath_clock(extendedBetaData, metaData, coefHorvath)
s
}
