% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_CpGs.R
\name{impute_missing_CpGs}
\alias{impute_missing_CpGs}
\title{Impute_missing_CpGs includes OSMI, impute.knn, methyLImp to impute missing values}
\usage{
impute_missing_CpGs(
  extendedBetaData,
  nRows = nRows,
  nCols = nCols,
  nRepeat,
  ncores
)
}
\arguments{
\item{extendedBetaData}{An expression data frame containing CpGs in rows and samples and  CpG position in columns}

\item{nRows}{A list or Number of rows to be selected}

\item{nCols}{A list or Number of columns to be selected}

\item{nRepeat}{A Number of runs}

\item{ncores}{Number of system cores to be used}
}
\value{
A list of data  frames containing, 1. the subset without missing values,
2. imputation records with impute.knn, 3. imputation records with methyLImp, 4. imputation records with OSMI
}
\description{
Impute_missing_CpGs includes OSMI, impute.knn, methyLImp to impute missing values
}
