% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMI.R
\name{OSMI}
\alias{OSMI}
\title{Missing values imputation on single or multiple samples from expression data}
\usage{
OSMI(extendedBetaData, positionColumnName, subData = NULL)
}
\arguments{
\item{extendedBetaData}{An expression data frame containing CpGs in rows and samples and  CpG position in columns.}

\item{positionColumnName}{A string names corresponding of the column with CpG positions}

\item{subData}{A data frame (for experimental purposes only); a sub data frame of extendedBetaData. The imputation is only
carried out for this data frame in order to reduce the computational effort. The nearest neighbour CpG search
is always based on extendedBetaData}
}
\value{
A list with two components named 'imputed' and 'info'.
'imputed'  a data frame, where missing values have been imputed
'info'    a data frame with some additional information about each imputation
(row, column, imputed value, distance in base pairs to  the nearest neighbour
if detectable, kind of imputation (mean or nearest neighbour))
}
\description{
OSMI imputes missing DNA methylation values of CpGs on a chromosome within the strand.
It replaces missing values by methylation values of nearest CpGs (in base pairs). It can be applied in single-sample applications.
}
\examples{
OSMI(extendedBetaData, "pos", subData)
}
\author{
author Lutz Leistritz
}
