% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMI_on_Array_density.R
\name{OSMI_on_array_density}
\alias{OSMI_on_array_density}
\title{Impact of the density of the expression data on OSMI}
\usage{
OSMI_on_array_density(extendedBetaData, ncores = ncores)
}
\arguments{
\item{ncores}{Number of system cores to be used}

\item{blood_download}{An expression data frame containing CpGs in rows and samples in columns}
}
\value{
list containin the result data frame and the plot save in p
and the length of the array used for the imputation
}
\description{
OSMI_on_array_density figures out the impact of the density of an methylation array on OSMI imputation.
The denser the array, the should be the performance of OSMI.
}
