% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perfomance_check}
\alias{perfomance_check}
\title{Evaluate the performance of OSMI}
\usage{
perfomance_check(nRows = nRows, nCols = nCols, nRepeat = nRepeat)
}
\arguments{
\item{nRows}{A list or Number of rows to be selected}

\item{nCols}{A list or Number of rows to be selected}

\item{nRepeat}{A Number of runs}
}
\value{
A table or a list of tables containing the means values from the repetitions
}
\description{
Evaluate the performance of OSMI
}
\examples{
perfomance_check(nRows=c(400, 1000,3000,5000 ),nCols=c(1,2,5,10,50, 100, 500, 1000, 1500),nRepeat=10)
}
