% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perfomance_plot}
\alias{perfomance_plot}
\title{Plot function to present the result in a graphic}
\usage{
perfomance_plot(tab_mean = tab_mean)
}
\arguments{
\item{tab_mean}{A data table with samples id on the columns and fucntion name on rows}
}
\value{
A list of plots
}
\description{
Plot function to present the result in a graphic
}
\examples{
perfomance_plot<-function(tab_mean=tab_mean)

}
