% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perfomance_plot.t}
\alias{perfomance_plot.t}
\title{A plot for the test on methyLImp 400 CpGs and 1000 Cpgs by 100 repeat}
\usage{
perfomance_plot.t(tab_mean = tab_mean)
}
\arguments{
\item{tab_mean}{A table with the means results from the imputation}
}
\value{
A list of plots
}
\description{
A plot for the test on methyLImp 400 CpGs and 1000 Cpgs by 100 repeat
}
