% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_CpGs.R
\name{sample_generator}
\alias{sample_generator}
\title{sample_generator subsets a large data set into a smaller and introduces missing values}
\usage{
sample_generator(extendedBetaData = extendedBetaData, rnum = rnum, cnum = cnum)
}
\arguments{
\item{extendedBetaData}{An complete cases data frame.}

\item{rnum}{Number of rows to be selected}

\item{cnum}{Nmber of columns to be selected}
}
\value{
A list of data frames containing in 1. the subset without missing values, 2. The  missing values, 3. The missng values possitions.
}
\description{
sample_generator first resamples a large data set into a smaller according to the requested number of rows and columns,
then adds 10 \% missing values to 90 \% of rows.
}
