% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_CpGs.R
\name{use_imputeknn}
\alias{use_imputeknn}
\title{Use the function impute.knn to impute missing values}
\usage{
use_imputeknn(datalist = datalist)
}
\arguments{
\item{datalist}{A list of data frames containing in 1. the subset without missing values, 2. The  missing values, 3. The missng values possitions.}
}
\value{
a list containing a data frame with records of the impution, and the execution time
}
\description{
Use the function impute.knn to impute missing values
}
