% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_CpGs.R
\name{use_osmi}
\alias{use_osmi}
\title{Use OSMI to impute missing values at a single and multiple subject levels.}
\usage{
use_osmi(extendedBetaData, subData, df_annotation, ncores)
}
\arguments{
\item{extendedBetaData}{An expression data frame containing CpGs in rows and samples and  CpG position in columns.}

\item{subData}{A data frame (for experimental purposes only); a sub data frame of extendedBetaData. The imputation is only}

\item{ncores}{Number of system cores to be used}
}
\value{
A list containing a data frame with records of the impution, and the execution time
}
\description{
Use OSMI to impute missing values at a single and multiple subject levels.
}
\examples{
use_osmi(extendedBetaData, subData,df_annotation,ncores)
}
