% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminateMissingsMulti.R
\name{eliminateMissingsMulti}
\alias{eliminateMissingsMulti}
\title{Eliminate Missing Values}
\usage{
eliminateMissingsMulti(dataFrame, max_cycles = Inf, delete_rows = Inf)
}
\arguments{
\item{dataFrame}{An data frame with missing values. For instance, an expression data frame containing CpGs in rows and samples in columns}

\item{max_cycles}{maximal cycle}

\item{delete_rows}{rows to be deleted}
}
\value{
list  a list of row and column names of the final complete matrix.
If parameters for early stopping have been specified, the final
matrix may still contain missing values.
}
\description{
The function applies a greedy algorithm to iteratively delete rows and columns of a matrix that contains missing values
to generate as large a complete matrix as possible.
}
\examples{
eliminateMissingsMulti(expression data frame)
}
\author{
Lutz Leistritz
Jena University Hospital, Jena, Germany
}
