% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_CpGs.R
\name{extendedBetaData_prep}
\alias{extendedBetaData_prep}
\title{extendedBetaData_prep splits an expression  data frame over chromosomes information into a list of data frames}
\usage{
extendedBetaData_prep(extendedBetaData)
}
\arguments{
\item{extendedBetaData}{An expression data frame containing CpGs in rows and samples and  CpG position in columns.}
}
\value{
A list of data frames in which CpGs were grouped depending on their position by chromosomes and strand
}
\description{
extendedBetaData_prep splits an expression  data frame over chromosomes information into a list of data frames
}
