% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_CpGs.R
\name{use_methylimp}
\alias{use_methylimp}
\title{Use the function methyLImp to impute missing values}
\usage{
use_methylimp(datalist = datalist)
}
\arguments{
\item{datalist}{A list of data frames containing in 1. the subset without missing values, 2. The  missing values, 3. The missng values possitions.}
}
\value{
A list containing a data frame with records of the imputation, and the execution time
}
\description{
Use the function methyLImp to impute missing values
}
\examples{
use_methylimp(x) ### x is the list resulted from sample_generator
}
