% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_CpGs.R
\name{use_osmi_island}
\alias{use_osmi_island}
\title{Use OSMI_island to impute missing values at a single and multiple subject levels.}
\usage{
use_osmi_island(extendedBetaData, ncores)
}
\arguments{
\item{extendedBetaData}{An expression data frame containing CpGs in rows and samples and  CpG position in columns.}

\item{ncores}{Number of system cores to be used}
}
\value{
A data frame with imputation scores
}
\description{
Use OSMI_island to impute missing values at a single and multiple subject levels.
}
